Get data from Maryland iMap

Get data from Maryland iMap including Computer Assisted Mass Appraisal (CAMA), parcel data, Maryland Inventory of Historic Properties data, and a variety of water data.

Usage

get_imap_data(
  location = NULL,
  nm = NULL,
  crs = getOption("mapmaryland.crs", default = 3857),
  ...
)

get_cama_data(
  location,
  type = "core",
  crs = getOption("mapmaryland.crs", default = 3857),
  ...
)

get_parcel_data(
  location,
  type = "boundaries",
  crs = getOption("mapmaryland.crs", default = 3857),
  block = FALSE,
  ...
)

format_parcel_data(
  data,
  imap_addr_cols = list(bldg_num = "strtnum", street_dir_prefix = "strtdir", street_name
    = "strtnam", street_suffix = "strttyp")
)

get_mht_data(
  location,
  type = "mihp",
  crs = getOption("mapmaryland.crs", default = 3857),
  ...
)

get_water_data(
  location,
  type = "streams",
  crs = getOption("mapmaryland.crs", default = 3857),
  ...
)

get_imap_url(nm = NULL)

Arguments

location

A sf, sfc, or bbox object (or other object convertible with sfext::as_bbox(). Required.

nm

Layer name identifier used to retrieve url from md_imap_index based on the snakecase "nm" column; Default: NULL.

crs

Coordinate reference system to return. Defaults to getOption("mapmaryland.crs", default = 3857).

...

Additional parameters passed to getdata::get_esri_data()

type

Default and supported options vary by function. See details.

block

If block is TRUE, pass parcel data to getdata::bind_block_col(). Defaults to FALSE.

data

Data to pass to getdata::bind_block_col(). format_parcel_data() only.

imap_addr_cols

Named list with address column names for input data. Passed to corresponding parameters in getdata::bind_block_col()

Details

Options for type parameter:

get_cama_data(): Default: "core"; options: "bldg" (detailed building characteristics), "core", "land", "subarea" (building subarea).

get_water_data(): Default: "streams", options: "streams" (Rivers and streams generalized), "streams detailed" (Rivers and streams generalized), "federal watersheds", "8 digit watersheds", "12 digit watersheds", "navigable waterways", "ponds" (Lakes and ponds)

get_parcel_data(): Default "boundaries", options: "boundaries" (Parcel boundaries), "points" (Parcel points)

get_mht_data(): Default: "mihp"; options: "mihp" (Maryland Inventory of Historic Properties), "nr" (National Register of Historic Places), "easements" (Preservation Easements held by the Maryland Historical Trust)

Helper functions:

format_parcel_data() is a wrapper for getdata::bind_block_col() used by get_parcel_data()

get_imap_url() is a look-up function for the md_imap_index using the short "nm" column as a index.