Use package data or the tigris package to get Maryland data from the U.S. Census Bureau

Get package data with Maryland counties, U.S. Congressional District, or Census designated places by name or GeoID or use the tigris package to download state data.

Usage

get_md_tigris(
  name = NULL,
  type = "counties",
  crs = 3857,
  erase_water = FALSE,
  ...
)

Arguments

name

Name matching a value in the namelsad, namelsad, or geoid columns of md_counties, md_census_places, md_congressional_districts, md_legislative_districts, or md_senate_districts, Default: NULL

type

Type of data to return, Default: "counties"; See details for supported options.

...

Additional parameters passed on to tigris functions.

Value

A simple feature object matching the type provided.

Details

Use mapmaryland packaged data or use tigris U.S. Census Bureau API:

Supported options that access data included with the mapmaryland package include "counties", "census places", "congressional districts", "legislative districts", and "water".

Options that access data by calling one of several tigris functions for downloading data from the U.S. Census Bureau API include "senate district", "tracts", "block groups", "blocks", "pumas", "voting districts". tigris functions that do not use a "state" parameter (e.g. tigris::zctas) are not supported by this function.