Maryland ArcGIS REST API Services Index

A non-comprehensive list of ArcGIS services maintained by public agencies, academic units, and nonprofit organizations in Maryland. Layers available at the service_url can be accessed using the esri2sf package.

Usage

md_arcgis_index

Format

A data frame with 95 rows and 13 variables:

name

character Source name

operator.abb

character Operator abbreviation

operator

character Operator name

city

character City name

county

character County name

state

character State name

state.abb

character State name abbreviation

geography

character Geographic scope; Options include citywide, countywide, regional (multi-county), statewide, or regional (multistate).

source_type

character Source type

notes

character Notes

public

logical Public data indicator

services_url

character Services URL

hosting

character Hosting type; Currently used as an indicator for ArcGIS Online hosted data with other services assumed to be self-hosted by the source or operator.

Details

Comments, additions, or corrections can be submitted using comments on this open Google Sheet: https://docs.google.com/spreadsheets/d/1c829bZdNqvbpoizulBU_XE5jVeNNck2kHkS-smpQ52s/edit?usp=sharing

Data last updated from Google Sheet 2024-09-18.