Maryland K-12 Schools (Public and Charter Schools)

Combined iMap FeatureLayer for traditional public schools and charter schools. The school number previously included in this dataset was incomplete due to the iMap data using a different school name than the MSDE data for a little less than half of schools in the state. As of February 2023, the school number column has been removed to ensure the data is accurate and to avoid confusion. For charter schools, the school_type column has been added based on the grade range. Data based on iMap data layers as of 2023 February 10 for public schools https://geodata.md.gov/imap/rest/services/Education/MD_EducationFacilities/FeatureServer/5 and charter schools https://geodata.md.gov/imap/rest/services/Education/MD_EducationFacilities/FeatureServer/6

Usage

md_schools

Format

A data frame with 1,427 rows and 13 variables:

school_name

School name

address

Street address

city

City name

state

State

zip

Zip code

grades

Grades

school_type

School type

grade_band

Grade band

psc_number

PSC number

lss_number

Local school system number as integer

county

County name

management_type

Management type

geometry

POINT geometry