Maryland Arts & Entertainment Districts

The Maryland State Arts Council sponsors Arts & Entertainment Districts as a way to stimulate the economy and improve quality of life. This legislation enables local jurisdictions, municipalities, counties, or a combination thereof, to apply for state designations for the Arts and Entertainment Districts within their boundaries and offer tax incentives as provided by law. More information from the Maryland State Arts Council: https://msac.org/programs/arts-entertainment-districts

Usage

md_arts_districts

Format

A data frame with 29 rows and 12 variables:

name

District name

desc

District description

org_name

Organization name

org_url

Organization website URL

org_phone

Organization phone number

org_address

Organization street address

org_city

Organization city

org_county

Organization county

zipcode

Organization zipcode

lon

Longitude

lat

Latitude

geometry

MULTIPOLYGON geometry with boundaries

Source

https://geodata.md.gov/imap/rest/services/BusinessEconomy/MD_IncentiveZones/FeatureServer/2